Extract Content API logic into separate class for clarity
